+++ src/main/java/de/maxhenkel/miningdimension/config/ClientConfig.java
@@ -1,13 +1,13 @@
 package de.maxhenkel.miningdimension.config;
 
 import de.maxhenkel.corelib.config.ConfigBase;
-import net.minecraftforge.common.ForgeConfigSpec;
+import net.neoforged.neoforge.common.ModConfigSpec;
 
 public class ClientConfig extends ConfigBase {
 
-    public final ForgeConfigSpec.BooleanValue showCustomWorldWarning;
+    public final ModConfigSpec.BooleanValue showCustomWorldWarning;
 
-    public ClientConfig(ForgeConfigSpec.Builder builder) {
+    public ClientConfig(ModConfigSpec.Builder builder) {
         super(builder);
         showCustomWorldWarning = builder
                 .comment("If the game should show the custom world warning when loading a world")