+++ src/main/java/de/maxhenkel/miningdimension/dimension/MiningDimensionTeleporter.java
@@ -14,8 +14,7 @@ import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.chunk.LevelChunk;
 import net.minecraft.world.phys.Vec3;
-import net.minecraftforge.common.util.ITeleporter;
-
+import net.neoforged.neoforge.common.util.ITeleporter;
 import javax.annotation.Nullable;
 import java.util.Map;
 import java.util.Objects;